requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
email-validator==2.1.0
openpyxl==3.1.2
python-dateutil==2.8.2
Flask-Cors==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
